Rename package from misleading "api" to more basic "model"
