+++ app/src/main/java/pl/borys/decide/common/KodeinProvider.kt
@@ -2,7 +2,7 @@ package pl.borys.decide.common
 
 import android.support.annotation.VisibleForTesting
 import org.kodein.di.Kodein
-import pl.borys.decide.usecase.vote.api.fakeVoteModule
+import pl.borys.decide.usecase.vote.model.fakeVoteModule
 
 object KodeinProvider {
     var kodeinInstance = Kodein {
